dsns/dsns/src/main/java/models/User.java
package models;

import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Data;

@Entity
@Table(name = "User")
@Data
public class User {
    @Id
    private Long id;
    private String username;
    private String password;
    private Integer phone;

    public User(Long id, String username, String password) {
        this.id = id;
        this.username = username;
        this.password = password;
        this.phone = phone;
    }

    public User() {

    }

    public User(String johnDoe, String mail) {
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getId() {
        return id;
    }
}
